refactor(new file): broadcast_shape +  can_broadcast.
