+++ libs/nox/src/array/mod.rs
@@ -25,9 +25,11 @@ use faer::{
 };
 use smallvec::SmallVec;
 
+mod broadcast;
 mod dynamic;
 mod repr;
 mod view;
+pub use broadcast::*;
 pub use repr::*;
 pub use view::*;
 
@@ -1309,41 +1311,6 @@ impl<D1: Dim, D2: Dim> MappableDim for (D1, D2) {
         D: Dim;
 
     type ElemDim = D2;
-}
-
-pub fn broadcast_shape(left: &[usize], right: &[usize]) -> Result<SmallVec<[usize; 4]>, Error> {
-    let rank = left.len().max(right.len());
-    let mut shape = SmallVec::with_capacity(rank);
-
-    for axis in 0..rank {
-        let left_axis = axis
-            .checked_sub(rank - left.len())
-            .map(|i| left[i])
-            .unwrap_or(1);
-        let right_axis = axis
-            .checked_sub(rank - right.len())
-            .map(|i| right[i])
-            .unwrap_or(1);
-
-        if left_axis == right_axis {
-            shape.push(left_axis);
-        } else if left_axis == 1 {
-            shape.push(right_axis);
-        } else if right_axis == 1 {
-            shape.push(left_axis);
-        } else {
-            return Err(Error::BroadcastShapeMismatch {
-                left: left.to_vec(),
-                right: right.to_vec(),
-            });
-        }
-    }
-
-    Ok(shape)
-}
-
-pub fn can_broadcast(left: &[usize], right: &[usize]) -> bool {
-    broadcast_shape(left, right).is_ok()
 }
 
 pub(crate) fn cobroadcast_dims(output: &mut [usize], other: &[usize]) -> bool {
